requirements.txt
googlemaps
Pillow
telepot
datetime
openpyxl